+++ src/main/java/domain/amount/YearlyInstallmentInvestmentAmount.java
@@ -16,7 +16,7 @@ public class YearlyInstallmentInvestmentAmount implements InstallmentInvestmentA
 
 	@Override
 	public double calAnnualInterest(InterestRate interestRate) {
-		return interestRate.getAnnualInterest(amount);
+		return interestRate.getAnnualInterest(amount).doubleValue();
 	}
 
 	@Override
